Close #9639: autodoc: Support asynchronous generator functions
